first step to link UI

+++ src/main/java/com/codetutr/bo/mapper/CustomerMapper.java
@@ -9,7 +9,7 @@ import com.codetutr.bo.Customer;
 public class CustomerMapper  implements RowMapper<Customer> {
     public Customer mapRow(ResultSet rs, int rowNum) throws SQLException {
         Customer customer = new Customer();
-        customer.setId(rs.getLong("id"));
+        customer.setId(rs.getInt("id"));
         customer.setName(rs.getString("name").trim());
         customer.setAddress(rs.getString("address").trim());
         customer.setMobile(rs.getLong("mobile"));
